Structure code in logically separated submodules
