+++ src/akkudoktoreos/devices/battery.py
