+++ src/akkudoktoreos/devices/generic.py
