+++ src/akkudoktoreos/utils/utils.py
@@ -1,5 +1,8 @@
 import datetime
+import json
 import zoneinfo
+
+import numpy as np
 
 
 # currently unused
@@ -14,6 +17,27 @@ def ist_dst_wechsel(tag: datetime.datetime, timezone="Europe/Berlin") -> bool:
     dst_change = current_day.replace(tzinfo=tz).dst() != next_day.replace(tzinfo=tz).dst()
 
     return dst_change
+
+
+class NumpyEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, np.ndarray):
+            return obj.tolist()  # Convert NumPy arrays to lists
+        if isinstance(obj, np.generic):
+            return obj.item()  # Convert NumPy scalars to native Python types
+        return super(NumpyEncoder, self).default(obj)
+
+    @staticmethod
+    def dumps(data):
+        """Static method to serialize a Python object into a JSON string using NumpyEncoder.
+
+        Args:
+            data: The Python object to serialize.
+
+        Returns:
+            str: A JSON string representation of the object.
+        """
+        return json.dumps(data, cls=NumpyEncoder)
 
 
 # # Example usage
